requirements.txt
yacs==0.1.8
numpy==1.19.5
torch==1.7.1
torchvision==0.8.2
matplotlib==3.3.3
tabulate==0.8.7
tensorflow==2.4.0
tensorboard==2.4.0
pillow==8.1.0
tqdm==4.56.0
opencv-python~=4.5
jupyter==1.0.0
pandas==1.1.0
